Various changes. (read more)
This is a kind of big commit contains a first test of our "ChestLogic". In it's current state, the bot is able to open chests (but not start unlocking any others). There are probably still false-positives though, which may or may not cause the bot to get stuck (until it forces a restart).

We also added support for 1080p resolutions (even though not all images are included yet!). This means you could actually bot on your phone (if you got a phone with a 1080p screen) using ADB.